Break components
