+++ packages/builder/src/pages/builder/portal/settings/auth/index.svelte
@@ -27,11 +27,11 @@
   import { onMount } from "svelte"
   import { API } from "api"
   import { organisation, admin, licensing } from "stores/portal"
+  import Scim from "./scim.svelte"
 
   const ConfigTypes = {
     Google: "google",
     OIDC: "oidc",
-    SCIM: "scim",
   }
 
   const HasSpacesRegex = /[\\"\s]/
@@ -161,8 +161,6 @@
     providers.oidc?.config?.configs[0].clientSecret
   )
 
-  $: scimEnabled = true
-
   const onFileSelected = e => {
     let fileName = e.target.files[0].name
     image = e.target.files[0]
@@ -256,19 +254,6 @@
     originalGoogleDoc = cloneDeep(providers.google)
   }
 
-  async function saveSCIM() {
-    try {
-      const res = await saveConfig({
-        type: "scim",
-        enabled: scimEnabled,
-      })
-      notifications.success(`Settings saved`)
-    } catch (e) {
-      notifications.error(e.message)
-      return
-    }
-  }
-
   let defaultScopes = ["profile", "email", "offline_access"]
 
   const refreshScopes = idx => {
@@ -358,14 +343,6 @@
     } else {
       originalOidcDoc = cloneDeep(oidcDoc)
       providers.oidc = oidcDoc
-    }
-
-    try {
-      const scimConfig = await API.getConfig(ConfigTypes.SCIM)
-      scimEnabled = scimConfig?.enabled
-    } catch (error) {
-      console.error(error)
-      notifications.error("Error fetching SCIM config")
     }
   })
 </script>
@@ -638,20 +615,7 @@
     </div>
 
     <Divider />
-    <Layout gap="XS" noPadding>
-      <div class="provider-title">
-        <Heading size="S">SCIM</Heading>
-      </div>
-      <Body size="S">Sync users with your identity provider.</Body>
-      <div class="form-row">
-        <Label size="L">Activated</Label>
-        <Toggle text="" bind:value={scimEnabled} />
-      </div>
-    </Layout>
-
-    <div>
-      <Button cta on:click={saveSCIM}>Save</Button>
-    </div>
+    <Scim />
   {/if}
 </Layout>
